change pointer to uint64 so HashMap doesn't complain

+++ src/ref_db.h
@@ -18,7 +18,7 @@ private:
     };
 
     Mutex* mut;
-    HashMap<uintptr_t, RefIndex> ref_map;
+    HashMap<uint64_t , RefIndex> ref_map;
     List<RefIndex> freeIds;
 
 public:
